+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -212,6 +212,8 @@ object Constants {
         const val BUILD_SEARCHABLE_OPTIONS = "buildSearchableOptions"
         const val CLEAN_SANDBOX = "cleanSandbox"
         const val COMPOSED_JAR = "composedJar"
+        const val DUMP_BUNDLED_MODULES = "dumpBundledModules"
+        const val DUMP_BUNDLED_PLUGINS = "dumpBundledPlugins"
         const val DUMP_PRODUCTS_RELEASES = "dumpProductsReleases"
         const val GENERATE_LEXER = "generateLexer"
         const val GENERATE_MANIFEST = "generateManifest"