+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -616,6 +616,8 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
         listOf(
             InitializeIntelliJPlatformPluginTask,
             DumpProductsReleasesTask,
+            DumpBundledPluginsTask,
+            DumpBundledModulesTask,
             PrintBundledModulesTask,
             PrintBundledPluginsTask,
             PrintProductsReleasesTask,